Rename fifty -> n points message template
